Move test to new class
